Refactor into platform load
